+++ parglare/glr.py
@@ -71,7 +71,7 @@ class GLRParser(Parser):
         """
         pass
 
-    def parse(self, input_str, position=0, file_name=None, extra=None, clear=True):
+    def parse(self, input_str, position=0, file_name=None, extra=None):
         """
         Parses the given input string.
         Args:
@@ -80,7 +80,6 @@ class GLRParser(Parser):
             file_name(str): File name if applicable. Used in error reporting.
             extra: An object that keeps custom parsing state. If not given
                 initialized to dict.
-            clear: should we clear transient state after parsing.
         """
 
         if self.debug:
@@ -96,7 +95,6 @@ class GLRParser(Parser):
         self.input_str = input_str
         self.file_name = file_name
         self.extra = {} if extra is None else extra
-        self.clear = clear
 
         # Error reporting and recovery
         self.errors = []
@@ -162,12 +160,12 @@ class GLRParser(Parser):
             if self.debug:
                 a_print(f"*** {forest.solutions} successful parse(s).")
 
-            if self.clear:
+            if self.clear_transient:
                 self._remove_transient_state()
             return forest
         else:
             # Report error
-            if self.clear:
+            if self.clear_transient:
                 self._remove_transient_state()
             error = self.errors[-1]
             del self.errors